+++ vector.h
@@ -3,12 +3,12 @@
 
 typedef struct s_vector
 {
-	double x;
-	double y;
-	double z;
+	float x;
+	float y;
+	float z;
 } t_vector;
 
-t_vector	*vector_new(double x, double y, double z);
+t_vector	*vector_new(float x, float y, float z);
 t_vector	*vector_subtract(t_vector *v1, t_vector *v2);
 float vector_length(t_vector *v);
 void vector_normalize(t_vector *v);
